[All] Resolve windows warning

+++ client/ConcreteObject.cpp
@@ -14,5 +14,5 @@ ConcreteObject::~ConcreteObject()
 void			ConcreteObject::draw(double time)
 {
 	if (this->_sprite)
-		this->_sprite->draw(this->_x, this->_y, time);
+		this->_sprite->draw(static_cast<int>(this->_x), static_cast<int>(this->_y), time);
 }
